refactor: rename solution 2 ref poc files
