+++ packages/date-picker/src/Filter/components/FilterTriggerButton/FilterTriggerButton.tsx
@@ -4,7 +4,7 @@ import { Icon } from "@kaizen/component-library"
 import chevronDown from "@kaizen/component-library/icons/chevron-down.icon.svg"
 import chevronUp from "@kaizen/component-library/icons/chevron-up.icon.svg"
 import { isRefObject } from "../../../utils/isRefObject"
-import { FilterRef } from "../../FilterSolution2"
+import { FilterRef } from "../../FilterSolution2Ref"
 import { useFilterContext } from "../../context/useFilterContext"
 import { FilterBaseButton, FilterBaseButtonProps } from "../FilterBaseButton"
 import styles from "./FilterTriggerButton.module.scss"
